+++ aws-test/tests/aws_api_gatewayv2_route/variables.tf
@@ -1,0 +1,82 @@
+
+variable "resource_name" {
+  type    = string
+  default = "turbot-test-20200125-create-update"
+  description = "Name of the resource used throughout the test."
+}
+
+variable "aws_profile" {
+  type    = string
+  default = "default"
+  description = "AWS credentials profile used for the test. Default is to use the default profile."
+}
+
+variable "aws_region" {
+  type    = string
+  default = "us-east-1"
+  description = "AWS region used for the test. Does not work with default region in config, so must be defined here."
+}
+
+variable "aws_region_alternate" {
+  type    = string
+  default = "us-east-2"
+  description = "Alternate AWS region used for tests that require two regions (e.g. DynamoDB global tables)."
+}
+
+provider "aws" {
+  profile = var.aws_profile
+  region  = var.aws_region
+}
+
+provider "aws" {
+  alias   = "alternate"
+  profile = var.aws_profile
+  region  = var.aws_region_alternate
+}
+
+data "aws_partition" "current" {}
+data "aws_caller_identity" "current" {}
+data "aws_region" "primary" {}
+data "aws_region" "alternate" {
+  provider = aws.alternate
+}
+
+data "null_data_source" "resource" {
+  inputs = {
+    scope = "arn:${data.aws_partition.current.partition}:::${data.aws_caller_identity.current.account_id}"
+  }
+}
+
+resource "aws_apigatewayv2_api" "named_test_resource" {
+  name = var.resource_name
+  protocol_type = "HTTP"
+  tags = {
+    name = var.resource_name
+  }
+}
+
+
+resource "aws_apigatewayv2_route" "named_test_resource" {
+  api_id    = aws_apigatewayv2_api.named_test_resource.id
+  route_key = "$default"
+}
+
+output "resource_aka" {
+  value = "arn:${data.aws_partition.current.partition}:apigateway:${var.aws_region}::/apis/${aws_apigatewayv2_api.named_test_resource.id}/routes/${aws_apigatewayv2_route.named_test_resource.id}"
+}
+
+output "resource_id" {
+  value = aws_apigatewayv2_route.named_test_resource.id
+}
+
+output "api_id" {
+  value = aws_apigatewayv2_api.named_test_resource.id
+}
+
+output "api_endpoint" {
+  value = aws_apigatewayv2_api.named_test_resource.api_endpoint
+}
+
+output "resource_name" {
+  value = var.resource_name
+}